requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
openai==1.3.9
httpx==0.25.2
python-dotenv==1.0.0
gevent==23.9.1
gevent-websocket==0.10.1
requests==2.31.0
